Handle Ads Issuer Type in Cache
In Ads, issuer types have been made unique by appending a suffix in the
form of `_4`, where 4 is the issuer cohort. However, ads-serve does not
include this suffix in its requests (correctly). Until ads-serve can
integrate using public keys instead of name/cohort composites, the
challenge-bypass cache will account for the potential absence of the
suffix in fetches and ensure that the cache is populated using the
issuer_type from the database, rather than the one included in the
client request.