fix: handle unknown topic etc.
Refactor processMessagesIntoBatchPipeline to ensure that it closes the message
channel to signal that the message processing is ready including the case when
the message topic does not match any of configured topics. For that move all
processing of the message including the search for topic processors to
runMessageProcessor() groutine and use defer to close the channel. That required
to change the message channel to be of struct{} type and store the error in the
message itself so a simple `defer close(msg.done)` can be used to cover both
normal and error cases.

Provide unit tests for processMessagesIntoBatchPipeline that cover success and
error paths.

Bound batchPipeline capacity by the number of CPU cores to avoid running too
many CPU-intensive tasks in parallel.

Closes #702
Closes #703
Closes #704