Fork API and Kafka Execution
Rather than running Kafka and the API at the same time, run only Kafka
when Kafka is called for and only the API when the API is needed. This
branch also includes a refactor of the logging setup for Kafka so that
we can better assess the success of these changes.