Add Integration Tests
- Kafka token issuance
- Kafka token redemption
- V1 HTTP Issuer GET endpoint
- V1 HTTP Token Redemption POST endpoint
- V3 HTTP Token Redemption POST endpoint

This checks the entire token redemption flow for all known production
use cases as well as checks that duplicate redemtion fails. There are
some cases with time limited issuers where a token is issued for a
future issuer whose redemption fails. These tokens are tested and
verified to fail as expected.